Add error message for empty submission directory
The code was misbehaving and printing IndexError
messages with an empty directory of student submissions
when "Start Grading" was pressed. The traceback was:

```
Exception in Tkinter callback
Traceback (most recent call last):
  File "C:\Users\nbowd\anaconda3\envs\csvgrader\Lib\tkinter\__init__.py", line 1967, in __call__
    return self.func(*args)
           ^^^^^^^^^^^^^^^^
  File "C:\Users\nbowd\Documents\1 Python Scripts\csvgrade\src\csvgrader\gui\mainWindow.py", line 520, in enableGrading
    self.currentStudent = self.grader.getCurrentStudent()
                          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\nbowd\Documents\1 Python Scripts\csvgrade\src\csvgrader\Grader\grader.py", line 177, in getCurrentStudent
    netID = self.studentsSubmitted[self.currentStudent]
            ~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^
IndexError: list index out of range
```

And it was because the studentsSubmitted list
was empty. Thus added a check in `Grader.GenSubmitList()`
to ensure the number of students (list length) is not zero.