Add scrollbar to rubric column on grader tab
Adding a scrollbar is one of the feature goals
noted on the readme. Based on a Google AI result*,
the scrollbar cant be added to a Frame, but if
a Frame is nested in a Canvas, then the scrollbar
can be added to the Canvas. I thus did some nesting
of a sub frame and canvas in the original rubric frame
element.

This nesting was the easiest/quickest way to adapt
the example code, making the Canvas and scrollbar
children of the original frame. It might be cleaner
to make them children of the tab object, but I
was not sure how to put both onto the grid.

*Google Search: "tkinter add scrollbar to frame"