Fix Groups CSV read and write on Windows (cross-platform)
On Windows, the groups.csv was being written out and
gave empty lines between each name, which then could
not be read back in by the program. The output might
look like:

```
NetID,Group

student1,1

student2,1

```

So then `Groups.importGroups()` was failing with
IndexError at line 55 because the index 1 would
be out-of-bounds
```
for line in reader:
    self.addStudent(int(line[1]), line[0])
```

Per the documentation of csv.reader and csv.writer,
the cross-platform solution is to open files and disable
automatic line-ending conversion, because the csv module
will handle it internally, the issue arises if both
`open()` and `csv` are handling line endings, then
extra carriage returns get inserted on windows.

See:
https://docs.python.org/3/library/csv.html#csv.writer
https://stackoverflow.com/a/3191811