Feature: Remember program state from session to session.
I was finding it annoying to have to search through my
whole directory structure each time I started the program,
especially when just testing out programming changes.

I changed how some of the variables are assigned: the
gradebook path, submissions folder, and groups path
are now defined earlier based on defaults or by reading
in values from a memory.json file. The initial state
of the checkboxes is also read in.

Selecting a new directory with one of the buttons,
pushing "Start Grading", and pushing a new "Save State"
button all update the memory.json file with values
that the user currently has selected or inputted.

In this way, the last program state is remembered from
one session to the next. The memory.json file is not
committed to the repo, as it is different on a per-user
basis. It is stored in "src/csvgrader/memory.json"