Change installation behavior and required packages
Edit pyproject.toml to reflect only pandas as external
dependency, tkinter is a builtin python package.

Also add a GUI script which will be generated when
the package is pip-installed. Add a function to
call in csvgrader/__init__.py which implements this
GUI script. Also add a csvgrader/__main__.py,
so the module may be run even without install by typing
```
python -m path/to/module/csvgrader
```
Change imports in the files to relative imports,
based on "csvgrader" being the main package in "src".
This makes running the files in the above manner
work properly.

Change names of older Grader and MainWindow folders
to avoid confusion and test if they are accidentally
still used in the program. They do not seem to be.