Fix: need os.path.dirname not os.dirname
I had misnamed a function call in
`mainWindow.MainWindow.getPaths`, so
fixing that.

The reason the `getPaths` was changed for
this memory feature is so that it sensibly
opens the location presently remembered, rather
than the home directory, but if the location stored
is to a file, it needs to detect that and open the
parent directory instead.