Feat: selection titles, no overwrite names on cancel
Add titles to the file selection boxes, which can
be set to remind the user what they are looking
for (e.g. "Select Gradebook CSV").

Prevent unintended overwrites of the saved filename.
If the user opens file selection, but then cancels
out of it, the function returns None. It is not
helpful to set the remembered path to None or
an empty string, so that it will open again at
CSV Grader's working directory, thus add a check
to see if the filename has been passed before
updating the string variable.